+++ src/CosmicClock/CosmicClock.cpp
@@ -812,7 +812,7 @@ struct CosmicClockWidget : ModuleWidget {
 		addChild(createWidget<ScrewBlack>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 		addChild(createWidget<ScrewBlack>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 
-		eclipse::addHeader(this, 101.6f, "C O S M I C   C L O C K", "THE WHEEL OF HOURS");
+		eclipse::addHeader(this, 101.6f, "C O S M I C   C L O C K");
 
 		CosmicWheel* wheel = new CosmicWheel;
 		wheel->module = module;