+++ src/MoonPhaseDistortion/MoonPhaseDistortion.cpp
@@ -225,7 +225,7 @@ struct MoonPhaseDistortionWidget : ModuleWidget {
 		addChild(createWidget<ScrewBlack>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 		addChild(createWidget<ScrewBlack>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 
-		eclipse::addHeader(this, XC, "M O O N", "THE HOLLOW SUN");
+		eclipse::addHeader(this, XC, "M O O N");
 
 		MoonDisplay* display = new MoonDisplay;
 		display->module = module;