+++ src/Saros/Saros.cpp
@@ -385,7 +385,7 @@ struct SarosWidget : ModuleWidget {
 		addChild(createWidget<ScrewBlack>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 		addChild(createWidget<ScrewBlack>(Vec(box.size.x - 2 * RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 
-		eclipse::addHeader(this, XC, "S A R O S", "THE LONG RETURN");
+		eclipse::addHeader(this, XC, "S A R O S");
 
 		SarosDisplay* display = new SarosDisplay;
 		display->module = module;